Convert Twitter to a higher-order component
